# MusicManager_3/src/main/resources/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Music API
  version: 1.0.0
  description: API quản lý Playlist, Song, Authentication

servers:
  - url: http://localhost:8080
    description: Local server

tags:
  - name: Playlist
    description: Quản lý playlist
  - name: PlaylistSong
    description: Quản lý bài hát trong playlist
  - name: Song
    description: Quản lý bài hát
  - name: Authentication
    description: Đăng ký, đăng nhập

paths:
  /playlists:
    get:
      tags:
        - Playlist
      summary: Lấy tất cả playlists
      responses:
        '200':
          description: Danh sách playlists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Playlist'

    post:
      tags:
        - Playlist
      summary: Tạo playlist mới
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: playlist1
      responses:
        '200':
          description: Playlist đã tạo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Playlist'

  /playlists/{id}:
    get:
      tags:
        - Playlist
      summary: Tìm playlist theo ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Playlist theo ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Playlist'
    put:
      tags:
        - Playlist
      summary: Cập nhật playlist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: playlist1
      responses:
        '200':
          description: Playlist đã cập nhật
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Playlist'

    delete:
      tags:
        - Playlist
      summary: Xóa playlist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Xóa thành công

  /playlists/name/{name}:
    get:
      tags:
        - Playlist
      summary: Tìm playlist theo tên
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Danh sách playlists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Playlist'

  /playlistSong/playlists/{playlistId}/songs/{songId}:
    post:
      tags:
        - PlaylistSong
      summary: Thêm bài hát vào playlist
      parameters:
        - name: playlistId
          in: path
          required: true
          schema:
            type: integer
        - name: songId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Trả về PlaylistSong vừa thêm

    delete:
      tags:
        - PlaylistSong
      summary: Xóa bài hát khỏi playlist
      parameters:
        - name: playlistId
          in: path
          required: true
          schema:
            type: integer
        - name: songId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Xóa thành công

  /playlistSong/playlists/{playlistId}/songs:
    get:
      tags:
        - PlaylistSong
      summary: Lấy danh sách bài hát trong playlist
      parameters:
        - name: playlistId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Danh sách bài hát
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Song'

  /playlistSong/songs/{songId}/playlists:
    get:
      tags:
        - PlaylistSong
      summary: Lấy danh sách playlists chứa bài hát
      parameters:
        - name: songId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Danh sách playlists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Playlist'

  /songs:
    get:
      tags:
        - Song
      summary: Lấy tất cả bài hát
      responses:
        '200':
          description: Danh sách bài hát
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Song'
    post:
      tags:
        - Song
      summary: Tạo bài hát mới
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Song'
      responses:
        '200':
          description: Bài hát đã tạo

  /songs/{id}:
    get:
      tags:
        - Song
      summary: Tìm bài hát theo ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Bài hát theo ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Song'
    put:
      tags:
        - Song
      summary: Cập nhật bài hát
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Song'
      responses:
        '200':
          description: Bài hát đã cập nhật
    delete:
      tags:
        - Song
      summary: Xóa bài hát
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Xóa thành công

  /songs/name/{name}:
    get:
      tags:
        - Song
      summary: Tìm bài hát theo tên
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Danh sách bài hát
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Song'

  /auth/register:
    post:
      tags:
        - Authentication
      summary: Đăng ký người dùng mới
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: Người dùng đã đăng ký

  /auth/login:
    post:
      tags:
        - Authentication
      summary: Đăng nhập
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: JWT Token trả về
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JwtResponse'

components:
  schemas:
    Playlist:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string

    Song:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        artist:
          type: string
        genre:
          type: string
        filePath:
          type: string

    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        password:
          type: string
        role:
          type: string

    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string

    JwtResponse:
      type: object
      properties:
        token:
          type: string